add event receiver tests
